feat: improve autofixes
